Trekker Varsel ut i egen fil, det samme med testene som testet disse. Lagde feil-state og oppdaterer denne dersom behandlingstema + søknadsland er ugyldig kombinasjon. Denne hindrer bekreft og fortsett-trykk. Hentet ting direkte fra redux, istedenfor å sende det gjennom inngang.js. Fjerner tilsynelatende ubrukte felt i inngang.js, og stegvelger hvis jeg ikke finner andre bruk av dem.
