+++ src/felleskomponenter/stegvelger/Stegvelger.jsx
@@ -381,7 +381,6 @@ class Stegvelger extends Component {
       virksomheterIPerioden: props.arbeidsgivereIPerioden,
       avklartefakta: props.avklartefakta,
       begrunnelser: MKV.KTObjects.begrunnelser,
-      landkoder: props.landkoder,
       behandlingstype: props.oppsummering.behandlingstype,
       behandlingstema: props.oppsummering.behandlingstema,
       behandlingsstatus: props.oppsummering.behandlingsstatus,
@@ -417,7 +416,6 @@ class Stegvelger extends Component {
       hjemmebaser: props.hjemmebaser,
       harValgtNorskArbeidsgiver: props.harValgtNorskArbeidsgiver,
       medfolgendeBarn: props.medfolgendeBarn,
-      mottatteOpplysninger: props.mottatteOpplysninger,
       lagredeVirksomheter: props.lagredeVirksomheter,
       bestemmelser: props.bestemmelser,
       soknadsperiode: props.soknadsperiode,
@@ -624,8 +622,6 @@ Stegvelger.propTypes = {
   bestillAnmodningsperioder: PT.func.isRequired,
   harValgtNorskArbeidsgiver: PT.bool.isRequired,
   medfolgendeBarn: PT.array.isRequired,
-  mottatteOpplysninger: PT.object.isRequired,
-  landkoder: PT.arrayOf(MPT.Kodeverk).isRequired,
   lagredeVirksomheter: PT.array.isRequired,
   sakstype: PT.string,
   soknadsperiode: MPT.Soknadsperiode.isRequired,
@@ -727,7 +723,6 @@ const mapStateToProps = (state) => ({
   erArbeidEttLand: behandlingerSelectors.ErArbeidEttLand(state),
   harValgtNorskArbeidsgiver: flytSelectors.HarValgtNorskArbeidsgiverSelector(state),
   medfolgendeBarn: mottatteOpplysningerSelectors.MedfolgendeBarnSelector(state),
-  mottatteOpplysninger: mottatteOpplysningerSelectors.MottatteOpplysningerDataSelector(state),
   lagredeVirksomheter: oppsummertfaktaSelectors.VirksomhetIDerSelector(state),
   soknadsperiode: mottatteOpplysningerSelectors.PeriodeSelector(state),
   feilmeldinger: feiletResponsSelectors.FeilmeldingerSelector(state),
