+++ src/sider/eu_eøs/stegKomponenter/vurderingInngang/vurderingInngang.test.tsx
@@ -1,159 +1,38 @@
 import { ComponentProps } from "react";
 import { instance, mock } from "ts-mockito";
 
-import { render, screen } from "@testing-library/react";
+import { screen } from "@testing-library/react";
 
 import MKV from "../../../../melosyskodeverk";
 
-import { Varsler, VurderingInngang } from "./vurderingInngang";
-
-vi.mock("../../../../featuretoggle", () => ({
-  useFeatureToggle: () => true,
-}));
-
-describe("Varsler", () => {
-  const mockedProps = mock<ComponentProps<typeof Varsler>>();
-  let props = instance(mockedProps);
-  const konvensjonmelding =
-    "Husk at du må vurdere om inngangsvilkårene i konvensjonen med Storbritannia av 30. juni 2023, eller separasjonsavtalen av 28. januar 2020, er oppfylt.";
-
-  beforeEach(() => {
-    props = instance(mockedProps);
-    props.landkoder = ["DK"];
-  });
-
-  it("Viser melding om oppfyllte inngangsvilkår", () => {
-    props.oppfyllerInngangsvilkar = true;
-    props.inngangsvilkaar = {
-      vilkaar: MKV.Koder.vilkaar.FO_883_2004_INNGANGSVILKAAR,
-      oppfylt: true,
-      begrunnelseKoder: [],
-      begrunnelseFritekst: null,
-      begrunnelseFritekstEngelsk: null,
-    };
-
-    render(<Varsler {...props} />);
-
-    expect(screen.getByRole("listitem")).toHaveTextContent(
-      "Søknaden oppfyller inngangsvilkårene for EU/EØS-saker etter forordning 883/2004."
-    );
-  });
-
-  it("Viser feilmelding og hjelpetekst ved ikke oppfylte inngangsvilkår", () => {
-    props.oppfyllerInngangsvilkar = false;
-    props.inngangsvilkaar = {
-      ...props.inngangsvilkaar,
-      vilkaar: MKV.Koder.vilkaar.FO_883_2004_INNGANGSVILKAAR,
-      oppfylt: true,
-      begrunnelseFritekst: null,
-      begrunnelseFritekstEngelsk: null,
-      begrunnelseKoder: [
-        MKV.Koder.begrunnelser.inngangsvilkaar.MANGLER_STATSBORGERSKAP,
-        MKV.Koder.begrunnelser.inngangsvilkaar.TEKNISK_FEIL,
-      ],
-    };
-
-    render(<Varsler {...props} />);
-
-    expect(screen.getAllByRole("listitem")).toHaveLength(5);
-    expect(
-      screen.getByText("Søknaden oppfyller ikke inngangsvilkårene for EU/EØS-saker etter forordning 883/2004.")
-    ).toBeInTheDocument();
-    expect(screen.getByText(MKV.Terms.begrunnelser.inngangsvilkaar.MANGLER_STATSBORGERSKAP)).toBeInTheDocument();
-    expect(screen.getByText(MKV.Terms.begrunnelser.inngangsvilkaar.TEKNISK_FEIL)).toBeInTheDocument();
-    expect(screen.getByText("Du har to valg:")).toBeInTheDocument();
-    expect(screen.queryByText(konvensjonmelding)).toBeNull();
-  });
-
-  it("Viser feilmelding og hjelpetekst ved overstyrte inngangsvilkår", () => {
-    props.oppfyllerInngangsvilkar = true;
-    props.inngangsvilkaar = {
-      ...props.inngangsvilkaar,
-      vilkaar: MKV.Koder.vilkaar.FO_883_2004_INNGANGSVILKAAR,
-      oppfylt: true,
-      begrunnelseFritekst: null,
-      begrunnelseFritekstEngelsk: null,
-      begrunnelseKoder: [
-        MKV.Koder.begrunnelser.inngangsvilkaar.TEKNISK_FEIL,
-        MKV.Koder.begrunnelser.inngangsvilkaar.OVERSTYRT_AV_SAKSBEHANDLER,
-      ],
-    };
-
-    render(<Varsler {...props} />);
-
-    expect(screen.getAllByRole("listitem")).toHaveLength(4);
-    expect(
-      screen.getByText("Søknaden oppfyller ikke inngangsvilkårene for EU/EØS-saker etter forordning 883/2004.")
-    ).toBeInTheDocument();
-    expect(screen.getByText(MKV.Terms.begrunnelser.inngangsvilkaar.TEKNISK_FEIL)).toBeInTheDocument();
-    expect(screen.getByText("Du har to valg:")).toBeInTheDocument();
-    expect(screen.queryByText(konvensjonmelding)).toBeNull();
-  });
-
-  it("Viser konvensjonmelding ved overstyrte inngangsvilkår når land er GB og toggle er på og er utsendt behtema", () => {
-    props.oppfyllerInngangsvilkar = true;
-    props.inngangsvilkaar = {
-      ...props.inngangsvilkaar,
-      vilkaar: MKV.Koder.vilkaar.FO_883_2004_INNGANGSVILKAAR,
-      oppfylt: true,
-      begrunnelseFritekst: null,
-      begrunnelseFritekstEngelsk: null,
-      begrunnelseKoder: [
-        MKV.Koder.begrunnelser.inngangsvilkaar.TEKNISK_FEIL,
-        MKV.Koder.begrunnelser.inngangsvilkaar.OVERSTYRT_AV_SAKSBEHANDLER,
-      ],
-    };
-    props.behandlingstema = MKV.Koder.behandlinger.behandlingstema.UTSENDT_ARBEIDSTAKER;
-    props.landkoder = ["GB"];
-
-    render(<Varsler {...props} />);
-
-    expect(screen.getByText(konvensjonmelding)).toBeInTheDocument();
-  });
-
-  it("Viser feilmelding ved manglende inngangsvilkår", () => {
-    props.inngangsvilkaar = undefined;
-    props.behandlingHarPeriodeOgLand = true;
-
-    render(<Varsler {...props} />);
-
-    expect(screen.getByRole("listitem")).toHaveTextContent("Teknisk feil, finner ingen inngangsvilkår.");
-  });
-
-  it("Viser feilmelding ved manglende periode og land når det mangler", () => {
-    props.inngangsvilkaar = undefined;
-    props.behandlingHarPeriodeOgLand = false;
-
-    render(<Varsler {...props} />);
-
-    expect(screen.getByRole("listitem")).toHaveTextContent(
-      "Det mangler periode og/eller land. Fyll disse inn i sidemenyen nedenfor og oppdater registeropplysninger."
-    );
-  });
-
-  it("Viser feilmelding ved flere valgte land og ikke tema ARBEID_FLERE_LAND", async () => {
-    props.oppfyllerInngangsvilkar = true;
-    props.inngangsvilkaar = {
-      vilkaar: MKV.Koder.vilkaar.FO_883_2004_INNGANGSVILKAAR,
-      oppfylt: true,
-      begrunnelseKoder: [],
-      begrunnelseFritekst: null,
-      begrunnelseFritekstEngelsk: null,
-    };
-    props.behandlingstema = MKV.Koder.behandlinger.behandlingstema.UTSENDT_ARBEIDSTAKER;
-    props.landkoder = ["DK", "SE"];
-
-    render(<Varsler {...props} />);
-
-    expect(
-      await screen.findByText(
-        "Du har valgt et behandlingstema som kun tillater ett arbeidsland. Du må fjerne arbeidsland, eller endre behandlingstema for å kunne fatte vedtak."
-      )
-    ).toBeInTheDocument();
-  });
-});
+import { VurderingInngang } from "./vurderingInngang";
+import { renderWithProviders } from "../../../../ducks/test-utils/renderWithProviders";
 
 describe("VurderingInngang", () => {
+  const initialStore = (behandlingstema: string, landkoder: string[], flereLandUkjentHvilke: boolean) => ({
+    behandlinger: {
+      status: "",
+      data: {
+        oppsummering: {
+          behandlingstema: {
+            kode: behandlingstema,
+          },
+        },
+      },
+    },
+    mottatteOpplysninger: {
+      status: "",
+      data: {
+        data: {
+          soeknadsland: {
+            landkoder,
+            flereLandUkjentHvilke,
+          },
+        },
+      },
+    },
+  });
+
   const mockedProps = mock<ComponentProps<typeof VurderingInngang>>();
   let props = instance(mockedProps);
 
@@ -169,17 +48,12 @@ describe("VurderingInngang", () => {
         begrunnelseFritekstEngelsk: null,
       },
       oppfyllerInngangsvilkar: true,
-      behandlingID: 1,
-      hentVilkar: vi.fn(),
-      landkoder: ["DK"],
-      behandlingstema: MKV.Koder.behandlinger.behandlingstema.ARBEID_FLERE_LAND.kode,
-      behandlingHarPeriodeOgLand: true,
     };
   });
 
   describe("knapp for å gå videre i stegvelger", () => {
     it("er ikke disabled dersom redigerbart er true", () => {
-      render(<VurderingInngang {...props} />);
+      renderWithProviders(<VurderingInngang {...props} />);
 
       expect(screen.getByRole("button", { name: "Bekreft og fortsett" })).not.toBeDisabled();
     });
@@ -187,9 +61,45 @@ describe("VurderingInngang", () => {
     it("er disabled dersom redigerbart er false", () => {
       props.redigerbart = false;
 
-      render(<VurderingInngang {...props} />);
+      renderWithProviders(<VurderingInngang {...props} />);
 
       expect(screen.getByRole("button", { name: "Bekreft og fortsett" })).toBeDisabled();
     });
+  });
+
+  it("Viser feilmelding ved flere valgte land og ikke tema ARBEID_FLERE_LAND", async () => {
+    renderWithProviders(<VurderingInngang {...props} />, {
+      preloadedState: initialStore(MKV.Koder.behandlinger.behandlingstema.UTSENDT_ARBEIDSTAKER, ["DK", "SE"], false),
+    });
+
+    expect(
+      await screen.findByText(
+        "Du har valgt et behandlingstema som kun tillater ett arbeidsland. Du må fjerne land under “Periode og land” i sidemenyen eller endre behandlingstema."
+      )
+    ).toBeInTheDocument();
+  });
+
+  it("Viser feilmelding ved flereLandUkjentHvilke og ikke tema ARBEID_FLERE_LAND", async () => {
+    renderWithProviders(<VurderingInngang {...props} />, {
+      preloadedState: initialStore(MKV.Koder.behandlinger.behandlingstema.UTSENDT_ARBEIDSTAKER, [], true),
+    });
+
+    expect(
+      await screen.findByText(
+        "Du har valgt et behandlingstema som kun tillater ett arbeidsland. Du må fjerne land under “Periode og land” i sidemenyen eller endre behandlingstema."
+      )
+    ).toBeInTheDocument();
+  });
+
+  it("Viser feilmelding ved ett valgt land og tema ARBEID_FLERE_LAND", async () => {
+    renderWithProviders(<VurderingInngang {...props} />, {
+      preloadedState: initialStore(MKV.Koder.behandlinger.behandlingstema.ARBEID_FLERE_LAND, ["DK"], false),
+    });
+
+    expect(
+      await screen.findByText(
+        "Det er påkrevd med to eller flere land for valgt behandlingstema. Du må legge til land under “Periode og land” i sidemenyen eller endre behandlingstema."
+      )
+    ).toBeInTheDocument();
   });
 });
