+++ src/sider/eu_eøs/stegKomponenter/vurderingInngang/vurderingInngang.tsx
@@ -1,146 +1,20 @@
-import classNames from "classnames";
-import { connect, ConnectedProps } from "react-redux";
-import { RootState } from "AppTypes";
-import { ThunkDispatch } from "redux-thunk";
-import { Action } from "redux";
-
+import { useDispatch, useSelector } from "react-redux";
 import * as Nav from "../../../../navFrontend";
-import * as KV from "../../../../kodeverk";
-import * as Utils from "../../../../utils";
 import * as Api from "../../../../services/api";
 import * as Mui from "../../../../felleskomponenter/ui";
-
 import MKV, { MKVUtils } from "../../../../melosyskodeverk";
 import { mottatteOpplysningerSelectors } from "../../../../ducks/mottatteOpplysninger";
 import { behandlingerSelectors } from "../../../../ducks/behandlinger";
 import { vilkarOperations } from "../../../../ducks/vilkar";
-
 import "./vurderingInngang.css";
-import { useFeatureToggle } from "../../../../featuretoggle";
-import { MELOSYS_KONVENSJON_EFTA_LAND_OG_STORBRITANNIA } from "../../../../featuretoggle/toggleNavn";
+import { useEffect, useState } from "react";
+import Varsler from "./varsler";
 
-const { OVERSTYRT_AV_SAKSBEHANDLER } = MKV.Koder.begrunnelser.inngangsvilkaar;
-
-interface VarslerProps {
-  oppfyllerInngangsvilkar: boolean;
-  inngangsvilkaar: Api.Vilkar.Vilkaar | undefined;
-  landkoder: Array<string>;
-  behandlingstema: string;
-  behandlingHarPeriodeOgLand: boolean;
-}
-
-export const Varsler = ({
-  oppfyllerInngangsvilkar,
-  inngangsvilkaar,
-  landkoder,
-  behandlingstema,
-  behandlingHarPeriodeOgLand,
-}: VarslerProps) => {
-  const konvensjonStorbritanniaToggleEnabled = useFeatureToggle(MELOSYS_KONVENSJON_EFTA_LAND_OG_STORBRITANNIA);
-  const inngangsvilkaarBegrunnelseKoder = inngangsvilkaar?.begrunnelseKoder || [];
-
-  const inngangsvilkaarErOverstyrtAvSaksbehandler =
-    inngangsvilkaarBegrunnelseKoder.includes(OVERSTYRT_AV_SAKSBEHANDLER);
-
-  const inngangsvilkaarErOverstyrtEllerIkkeOppfylt =
-    inngangsvilkaarErOverstyrtAvSaksbehandler || !oppfyllerInngangsvilkar;
-
-  const inngangsvilkaarErOppfyltOgIkkeOverstyrt = oppfyllerInngangsvilkar && !inngangsvilkaarErOverstyrtAvSaksbehandler;
-
-  const oppfyllerInngangsvilkarCl = classNames({
-    liste__element: true,
-    "liste__element--oppfylt": inngangsvilkaarErOppfyltOgIkkeOverstyrt,
-    "liste__element--ikkeoppfylt": inngangsvilkaarErOverstyrtEllerIkkeOppfylt,
-  });
-
-  if (Utils._isEmpty(inngangsvilkaar)) {
-    if (!behandlingHarPeriodeOgLand) {
-      return (
-        <ul className="betingelser__liste">
-          <li className={oppfyllerInngangsvilkarCl}>
-            Det mangler periode og/eller land. Fyll disse inn i sidemenyen nedenfor og oppdater registeropplysninger.
-          </li>
-        </ul>
-      );
-    }
-    return (
-      <ul className="betingelser__liste">
-        <li className={oppfyllerInngangsvilkarCl}>Teknisk feil, finner ingen inngangsvilkår.</li>
-      </ul>
-    );
-  }
-
-  const oppfyltTekst = `Søknaden oppfyller${
-    inngangsvilkaarErOppfyltOgIkkeOverstyrt ? " " : " ikke "
-  }inngangsvilkårene for EU/EØS-saker etter forordning 883/2004.`;
-
-  const visbareInngangsvilkaarBegrunnelseKoder = inngangsvilkaarBegrunnelseKoder.filter(
-    (kode) => kode !== OVERSTYRT_AV_SAKSBEHANDLER
-  );
-
-  const flereSoknadslandEnnTillatt = landkoder.length > 1 && !MKVUtils.kanHaFlereSoknadsland(behandlingstema);
-
-  const visStorbritanniaKonvensjonTekst =
-    konvensjonStorbritanniaToggleEnabled &&
-    MKVUtils.erUtsendt(behandlingstema) &&
-    MKVUtils.enesteLandErStorbritannia(landkoder);
-
-  return (
-    <div className="vurderinginngang_eu_eos">
-      <ul className="betingelser__liste">
-        <li className={oppfyllerInngangsvilkarCl}>{oppfyltTekst}</li>
-        {inngangsvilkaarErOverstyrtEllerIkkeOppfylt &&
-          visbareInngangsvilkaarBegrunnelseKoder.map((begrunnelseKode) => (
-            <li key={begrunnelseKode} className={oppfyllerInngangsvilkarCl}>
-              {KV.kodeTilTerm(begrunnelseKode, MKV.KTObjects.begrunnelser.inngangsvilkaar)}
-            </li>
-          ))}
-      </ul>
-      {flereSoknadslandEnnTillatt && (
-        <Nav.Alert variant="warning">
-          Du har valgt et behandlingstema som kun tillater ett arbeidsland. Du må fjerne arbeidsland, eller endre
-          behandlingstema for å kunne fatte vedtak.
-        </Nav.Alert>
-      )}
-      {inngangsvilkaarErOverstyrtEllerIkkeOppfylt && (
-        <Nav.Alert variant="info">
-          {visStorbritanniaKonvensjonTekst && (
-            <p className="storbritannia-konv-tekst">
-              Husk at du må vurdere om inngangsvilkårene i konvensjonen med Storbritannia av 30. juni 2023, eller
-              separasjonsavtalen av 28. januar 2020, er oppfylt.
-            </p>
-          )}
-          Du har to valg:
-          <ul>
-            <li>Hvis inngangsvilkår ikke er oppfylt, må du avslutte saken fra behandlingsmenyen</li>
-            <li>Hvis inngangsvilkår er oppfylt, kan du fortsette behandlingen som normalt</li>
-          </ul>
-        </Nav.Alert>
-      )}
-    </div>
-  );
-};
-
-const mapStateToProps = (state: RootState) => ({
-  behandlingID: behandlingerSelectors.BehandlingIDSelector(state),
-  behandlingHarPeriodeOgLand: mottatteOpplysningerSelectors.HarPeriodeOgLandSelector(state),
-});
-
-const mapDispatchToProps = (dispatch: ThunkDispatch<RootState, unknown, Action>) => ({
-  hentVilkar: (behandlingID: number) => dispatch(vilkarOperations.hent(behandlingID)),
-});
-
-const connector = connect(mapStateToProps, mapDispatchToProps);
-type PropsFromRedux = ConnectedProps<typeof connector>;
-
-type VurderingInngangProps = PropsFromRedux & {
+type VurderingInngangProps = {
   bekreftOgFortsett: () => void;
   redigerbart: boolean;
   oppfyllerInngangsvilkar: boolean;
   inngangsvilkaar: Api.Vilkar.Vilkaar | undefined;
-  landkoder: Array<string>;
-  behandlingstema: string;
-  behandlingHarPeriodeOgLand: boolean;
 };
 
 export const VurderingInngang = ({
@@ -148,16 +22,37 @@ export const VurderingInngang = ({
   redigerbart,
   oppfyllerInngangsvilkar,
   inngangsvilkaar,
-  behandlingID,
-  hentVilkar,
-  landkoder,
-  behandlingstema,
-  behandlingHarPeriodeOgLand,
 }: VurderingInngangProps) => {
+  const dispatch = useDispatch();
+  const [feil, setFeil] = useState<string | undefined>(undefined);
+  const behandlingID = useSelector(behandlingerSelectors.BehandlingIDSelector);
+  const behandlingstema = useSelector(behandlingerSelectors.BehandlingstemaKodeSelector);
+  const behandlingHarPeriodeOgLand = useSelector(mottatteOpplysningerSelectors.HarPeriodeOgLandSelector);
+  const landkoder = useSelector(mottatteOpplysningerSelectors.SoknadslandkoderSelector);
+  const flereLandUkjentHvilke = useSelector(mottatteOpplysningerSelectors.SoknadslandFlereLandUkjentHvilkeSelector);
+
+  useEffect(() => {
+    const flereEnnEttLand = landkoder.length > 1 || flereLandUkjentHvilke;
+
+    if (flereEnnEttLand && !MKVUtils.kanHaFlereSoknadsland(behandlingstema)) {
+      setFeil(
+        "Du har valgt et behandlingstema som kun tillater ett arbeidsland. " +
+          "Du må fjerne land under “Periode og land” i sidemenyen eller endre behandlingstema."
+      );
+    } else if (!flereEnnEttLand && behandlingstema === MKV.Koder.behandlinger.behandlingstema.ARBEID_FLERE_LAND) {
+      setFeil(
+        "Det er påkrevd med to eller flere land for valgt behandlingstema. " +
+          "Du må legge til land under “Periode og land” i sidemenyen eller endre behandlingstema."
+      );
+    } else {
+      setFeil(undefined);
+    }
+  }, [landkoder, flereLandUkjentHvilke, behandlingstema]);
+
   const knappClickHandler = async () => {
     if (!oppfyllerInngangsvilkar) {
       await Api.Vilkar.overstyrInngangvilkaar(behandlingID);
-      await hentVilkar(behandlingID);
+      await dispatch(vilkarOperations.hent(behandlingID));
     }
 
     bekreftOgFortsett();
@@ -173,9 +68,10 @@ export const VurderingInngang = ({
         behandlingstema={behandlingstema}
         behandlingHarPeriodeOgLand={behandlingHarPeriodeOgLand}
       />
+      {feil && <Nav.Alert variant="error">{feil}</Nav.Alert>}
       <Mui.StegKnapper
         bekreftKnappProps={{
-          disabled: !redigerbart,
+          disabled: !redigerbart || Boolean(feil),
           onClick: knappClickHandler,
         }}
       />
@@ -183,4 +79,4 @@ export const VurderingInngang = ({
   );
 };
 
-export default connector(VurderingInngang);
+export default VurderingInngang;
